CF600E[AC]

+++ ssy/CF600E.cpp
@@ -1,19 +1,19 @@
 #include<bits/stdc++.h>
 using namespace std;
-const int N=1e5+10;
-int n,c[N];
-int head[N],nxt[2*N],ver[2*N],tot=1;
-void add(int u,int v){
+const long long N=1e5+10;
+long long n,c[N];
+long long head[N],nxt[2*N],ver[2*N],tot=1;
+void add(long long u,long long v){
     ver[++tot]=v;nxt[tot]=head[u];head[u]=tot;
 }
-int ans[N],mx[N];
-map<int,int>cnt[N];
-void dfs(int u,int p){
+long long ans[N],mx[N];
+map<long long,long long>cnt[N];
+void dfs(long long u,long long p){
     cnt[u][c[u]]=1;
     ans[u]=c[u];
     mx[u]=1;
-    for(int i=head[u];i;i=nxt[i]){
-        int v=ver[i];
+    for(long long i=head[u];i;i=nxt[i]){
+        long long v=ver[i];
         if(v==p)    continue;
         dfs(v,u);
         if(cnt[u].size()<cnt[v].size()){
@@ -31,18 +31,18 @@ void dfs(int u,int p){
     }
 }
 int main(){
-    scanf("%d",&n);
-    for(int i=1;i<=n;i++)
-        scanf("%d",&c[i]);
-    for(int i=1;i<n;i++){
-        int u,v;
-        scanf("%d%d",&u,&v);
+    scanf("%lld",&n);
+    for(long long i=1;i<=n;i++)
+        scanf("%lld",&c[i]);
+    for(long long i=1;i<n;i++){
+        long long u,v;
+        scanf("%lld%lld",&u,&v);
         add(u,v);
         add(v,u);
     }
     dfs(1,0);
-    for(int i=1;i<=n;i++)
-        printf("%d ",ans[i]);
+    for(long long i=1;i<=n;i++)
+        printf("%lld ",ans[i]);
     printf("\n");
     return 0;
 }
